requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
beautifulsoup4
langchain-openai
faiss_cpu
streamlit
langchain-ollama
langchain-groq
langchain-core
uvicorn
fastapi
langserve
